style: Refactor proposals to two-column tables for better readability (#44)
Currently, the table formatting for proposals is inconsistent: some use
a 1-column style, others use a 2-column style. The 2-column style is
more readable, as it more consistently separates the questions vs the
answers, without relying on visual flair such as **bold**.

I'm setting this PR to draft, as I reckon #43 is a precursor to this as
it combines the redundant questions at the bottom into a single question
(see that PR for details). If #43 is approved, I can reformat this PR to
reflect that change in the tables.

No substantial change to content has been made, only the form of the
tables (and a few minor typos, e.g., lots of uses of "Unkown" which
seems to be a mistake.)