Updates to the did:webvh DID Method Proposal to reflect v1.0 spec and latest information (#59)
Updates to the did:webvh DID Method standardization proposal to reflect
the recent change to v1.0 of the spec, and based on the latest
information. In addition to general updates, a link to a [DID Rubric
Evaluation](https://didwebvh.info/latest/did-rubric-evaluation/) has
been added.

It would be great if we could get this PR reviewed before the did:webvh
Deep Dive this coming Wednesday.

@peacekeeper, @jrayback, @bumblefudge

---------

Signed-off-by: Stephen Curran <swcurran@gmail.com>